chore: per reviewer feedback
