+++ src/helpers.ts
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
-// ** This file is automatically generated by gapic-generator-typescript. **
-// ** https://github.com/googleapis/gapic-generator-typescript **
-// ** All changes to this file may be overwritten. **
 import {
   googleProtobufValueFromObject,
   googleProtobufValueToObject,
@@ -26,7 +23,7 @@ interface ToValueFunction {
 
 interface FromValueFunction {
   new (): FromValueFunction;
-  fromValue(value: object): object | undefined; // TODO: Input needs to be Value; output needs to be Message
+  fromValue(value: object): object | undefined;
 }
 
 export function addToValue() {
