+++ src/value-converter.ts
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
-// ** This file is automatically generated by gapic-generator-typescript. **
-// ** https://github.com/googleapis/gapic-generator-typescript **
-// ** All changes to this file may be overwritten. **
 
 // TODO(): Remove this file once https://github.com/protobufjs/protobuf.js/pull/1495 is submitted.
 export function googleProtobufValueFromObject(
